chore: Add graphql module for better code base mgmt
