+++ changes/2669.misc.md
@@ -1,0 +1,1 @@
+Add `ai.backend.manager.models.graphql` module for better code base management.
