+++ src/ai/backend/manager/models/graphql/BUILD
@@ -1,0 +1,1 @@
+python_sources()
